refactor: http status codes 적용
