--- next-app/src/components/common/Grid/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Grid'
