+++ next-app/src/features/errors/errors.ts
@@ -1,5 +1,10 @@
+import httpStatus from 'http-status-codes'
+
 export const isErrorPage = (pathname: string) => {
-  return ['/404', '/500'].some((path) => {
+  return [
+    `/${httpStatus.NOT_FOUND}`, // 400
+    `/${httpStatus.INTERNAL_SERVER_ERROR}`, // 500
+  ].some((path) => {
     return pathname === path
   })
 }
